Build aarch64-linux-android on AArch64 Host
The build scripts now builds the llvm toolchain for AArch64 Host and
 Arm + AArch64 Target

To keep the produced toolchain compatible up with the current AOSP
codebase, its install directories have been named as per the naming
scheme of the default x86 llvm_toolchain variant.
This is expected to change in future.

We still require using prebuilt GCC libraries for building this
toolchain; Although this project will function exactly the same even
after building GCC on AArch64 is dealt with.

Signed-off-by: Gunwant Jain <therealgunwant@gmail.com>